+++ q1simulator/__init__.py
@@ -1,6 +1,7 @@
 __version__ = "0.8.1"
 
 from .q1simulator import Q1Simulator
+from .cluster import Cluster, ClusterModule
 
 from .q1viewer import plot_q1asm_file, plot_q1asm_files, PlotDef